Update Journal FFI to include GenesisController
This includes passing the location of the validator key to
the journal ffi so the key can be loaded and a Cylinder
Signer can be created.

When the journal is started, the genesis controller will
check if it needs to create a genesis block and do so before
starting the chain controller.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>